Add WhatsApp to VPN package blacklist
WhatsApp calls have not been working for a long time (see #590) and this can be annoying for many people. They might turn off Orbot completely and lose many of the privacy benefits and get angry and uninstall the app completely. For this reason, it makes sense to blacklist WhatsApp until a more comprehensive fix is found, as many people depend on it regularly.