Fix `defaultDisplay` deprecation in API 30
In this code, we first get the `DisplayManager` system service. Then we use the `getDisplay` method of `DisplayManager` to get the default display. Finally, we call the `getRealMetrics` method of Display to fill `displayMetrics` with data. This way, the code will work on both newer and older versions of Android.